Implement online editor (proof-of-concept)
This uses Monaco editor and the new KotlinJS target to implement an
online editor for Rhovas. This should be considered a proof-of-concept
and only covers the base editor window and output (as opposed to
including lists of examples, compiler options, or editor themes). These
features and better overall layout/experience will come later.

Basic syntax highlighted is provided by hijacking the basic-language
definition for Kotlin (since these names are built-in to Monaco) and
tweaking as needed. DSLs will syntax highlight as a string, but use
brace counting rather than indentation like the parser.